requirements.txt
requests
pandas
sqlalchemy
pymysql
python-dotenv
numpy
prophet